+++ apps/cli/cli_show.c
@@ -121,6 +121,7 @@ expand_dbvar(void   *h,
     char            *reason = NULL;
     char            *namespace = NULL;
     cvec            *nsc = NULL;
+    yang_stmt       *yp;
     
     if (argv == NULL || cvec_len(argv) != 2){
 	clicon_err(OE_PLUGIN, 0, "requires arguments: <db> <xmlkeyfmt>");
@@ -205,7 +206,12 @@ expand_dbvar(void   *h,
 	}
     if (xpath_vec_nsc(xcur, nsc, "%s", &xvec, &xlen, xpathcur) < 0) 
 	goto done;
-    bodystr0 = NULL; /* Assume sorted XML where duplicates are adjacent */
+    /* Loop for inserting into commands cvec. 
+     * Detect duplicates: for ordered-by system assume list is ordered, so you need
+     * just remember previous
+     * but for ordered-by system, check the whole list
+     */
+    bodystr0 = NULL;
     for (i = 0; i < xlen; i++) {
 	x = xvec[i];
 	if (xml_type(x) == CX_BODY)
@@ -214,10 +220,27 @@ expand_dbvar(void   *h,
 	    bodystr = xml_body(x);
 	if (bodystr == NULL)
 	    continue; /* no body, cornercase */
-	if (bodystr0 && strcmp(bodystr, bodystr0) == 0)
-	    continue; /* duplicate, assume sorted */
-	/* RFC3986 decode */
-	cvec_add_string(commands, NULL, bodystr);
+	if ((y = xml_spec(x)) != NULL &&
+	    (yp = yang_parent_get(y)) != NULL &&
+	    yang_keyword_get(yp) == Y_LIST &&
+	    yang_find(yp, Y_ORDERED_BY, "user") != NULL){
+	    /* Detect duplicates linearly in existing values */
+	    {
+		cg_var *cv = NULL;
+		while ((cv = cvec_each(commands, cv)) != NULL)
+		    if (strcmp(cv_string_get(cv), bodystr) == 0)
+			break;
+		if (cv == NULL)
+		    cvec_add_string(commands, NULL, bodystr);
+	    }
+	}
+	else{
+	    if (bodystr0 && strcmp(bodystr, bodystr0) == 0)
+		continue; /* duplicate, assume sorted */
+	    bodystr0 = bodystr;
+	    /* RFC3986 decode */
+	    cvec_add_string(commands, NULL, bodystr);
+	}
     }
  ok:
     retval = 0;